requirements.txt
requests==2.18.4
boto3==1.10.16
python_dateutil==2.8.1
